Data for the locations of pattern previews in StitchWorld 3 book
The JSON file was obtained by first extracting the pattern
previews with a bit of image processing (auto_crop.py), then
pattern numbers were OCRed from the cropped images, and
a bit of manual processing finished the job.

The script extract_patterns_from_pdf.py can be used to extract
the images from a PDF of the book. The boxes in the JSON file
work well for the PDF that can be downloaded on the Japanese
Brother website:
https://download.brother.com/welcome/doch102432/topical7pb01ja.pdf